feat: hook 추가, loader 추가
